Refactor AI level naming and settings layout
